requirements.txt
icecream
numpy
drake==1.22.0
matplotlib
opencv-python
colored
transforms3d
gym
pillow
meshcat
openai==0.28
ipdb
open3d
einops
dm-reverb
meshcat
tabulate
GPUtil
psutil
dotmap
zarr
transformers
sim_web_visualizer
sapien==2.2.2
mplib
webcolors
chardet
tensorboard
imageio
hydra-core==1.3.2
hydra-submitit-launcher==1.1.5
shortuuid
multimethod
chamferdist
sentence_transformers
wandb
easydict
timm
tqdm
scikit-learn==1.0.2
pickleshare==0.7.5
ninja==1.10.2.3
gdown
PyYAML==6.0
protobuf==3.19.4
termcolor==1.1.0
h5py==3.6.0
pyvista
setuptools==63.1.0
Cython
pandas
deepspeed
numba
ray
diffusers
noise
mplib==0.1.1
webcolors
ftfy
imageio[ffmpeg]